Fixed a bug preventing the user from inserting new sales on the database.

+++ syspharma/src/main/java/com/camaratapira/syspharma/repositories/ComprasRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.camaratapira.syspharma.entity.Compras;
 import com.camaratapira.syspharma.entity.Servidor;
 
-public interface ComprasRepository extends JpaRepository<Compras, Long>{
+public interface ComprasRepository extends JpaRepository<Compras, Integer>{
 	List<Compras> findByIdcompras(int idcompras);
 	List<Compras> findAllByIdservidor(final Servidor servidor);
